backend/requirements.txt
# FastAPI WhatsApp ChatBot Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.20
python-dotenv>=1.0.0
PyJWT>=2.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
twilio>=8.0.0
httpx>=0.25.0
requests>=2.30.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
websockets>=10.0
redis>=4.0.0
pytz>=2022.1
sqlalchemy>=2.0.0
anthropic>=0.18.0
psycopg2-binary>=2.9.0
